style(engine): complete license headers, fold eth_detail into detail, share test stubs
- Full Apache grant + @file/@brief on the five new files (N8)
- eth_detail folds into the shared bcos::engine::detail namespace; the
  blockFactory guard keeps std::invalid_argument as the legacy mirror
- kTargetNumber/kMaxEntries renamed to the c_ convention
- Stale comments no longer cite OpEngineReviewFixTest / a nonexistent S3
- EthServiceStubs.h now owns the storage/checkpoint/executor/scheduler
  stubs shared by the parity and transactional suites (EngineServiceTest
  keeps its copy until the #5547 side consolidates)
