+++ engine/bcos-engine/EthEngineService.h
@@ -1,6 +1,20 @@
 /**
  *  Copyright (C) 2026 FISCO BCOS.
  *  SPDX-License-Identifier: Apache-2.0
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ * @file EthEngineService.h
+ * @brief Ethereum Engine API service beside the legacy EngineServiceImpl
  */
 
 #pragma once
@@ -43,7 +57,7 @@
 namespace bcos::engine
 {
 
-namespace eth_detail
+namespace detail
 {
 template <class Guard, class ArtifactsMap>
 void commitRetainedPayload(Guard& guard, ArtifactsMap& artifacts, PayloadID const& payloadId,
@@ -63,7 +77,7 @@ void commitRetainedPayload(Guard& guard, ArtifactsMap& artifacts, PayloadID cons
         throw;
     }
 }
-}  // namespace eth_detail
+}  // namespace detail
 
 template <class ViewType>
 struct EthPayloadArtifacts
