+++ engine/bcos-engine/EthEngineService.inl
@@ -1,6 +1,20 @@
 /**
  *  Copyright (C) 2026 FISCO BCOS.
  *  SPDX-License-Identifier: Apache-2.0
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ * @file EthEngineService.inl
+ * @brief Ethereum Engine API service implementation (template bodies)
  */
 
 #pragma once
@@ -357,7 +371,7 @@ EthEngineService<MemPoolType, GlobalStateStorageType, ExecutorType, SchedulerTyp
     {
         auto guard = m_tracker.lockExclusive();
         // Keep the locally built body (executed at FCU). Do not rewrite from the CL request.
-        eth_detail::commitRetainedPayload(
+        detail::commitRetainedPayload(
             guard, m_artifacts, payloadId, cached->executionPayload.blockHash, cached);
     }
