+++ engine/test/unittests/engine/EthServiceStubs.h
@@ -1,0 +1,109 @@
+/**
+ *  Copyright (C) 2026 FISCO BCOS.
+ *  SPDX-License-Identifier: Apache-2.0
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ * @file EthServiceStubs.h
+ * @brief Shared test stubs for the EthEngineService suites (storage stack,
+ *        checkpoint backend, no-op executor/scheduler). One definition so the
+ *        suites cannot drift apart.
+ */
+
+#pragma once
+
+#include <bcos-framework/ledger/LedgerConfig.h>
+#include <bcos-framework/protocol/BlockHeader.h>
+#include <bcos-framework/protocol/Transaction.h>
+#include <bcos-framework/protocol/TransactionReceipt.h>
+#include <bcos-framework/storage2/MemoryStorage.h>
+#include <bcos-framework/storage2/MultiLayerStorage.h>
+#include <bcos-framework/transaction-executor/StateKey.h>
+#include <bcos-task/Task.h>
+#include <bcos-utilities/Common.h>
+#include <evmc/evmc.h>
+#include <range/v3/range/concepts.hpp>
+
+#include <cstdlib>
+#include <functional>
+#include <optional>
+
+namespace bcos::engine::eth_test
+{
+
+using RealGlobalStateMutableStorage = bcos::storage2::memory_storage::MemoryStorage<
+    bcos::executor_v1::StateKey, bcos::executor_v1::StateValue,
+    bcos::storage2::memory_storage::Attribute(bcos::storage2::memory_storage::ORDERED |
+                                              bcos::storage2::memory_storage::LOGICAL_DELETION)>;
+using RealGlobalStateBackendStorage =
+    bcos::storage2::memory_storage::MemoryStorage<bcos::executor_v1::StateKey,
+        bcos::executor_v1::StateValue,
+        bcos::storage2::memory_storage::Attribute(
+            bcos::storage2::memory_storage::ORDERED | bcos::storage2::memory_storage::CONCURRENT),
+        std::hash<bcos::executor_v1::StateKey>>;
+
+template <class Key, class Value, bcos::storage2::ReadWriteStorage<Key, Value> Storage>
+struct TrivialCheckpointStorage
+{
+    using CheckpointName = bcos::h256;
+    Storage& m_storage;
+    explicit TrivialCheckpointStorage(Storage& s) : m_storage(s) {}
+    Storage& open() { return m_storage; }
+    [[noreturn]] Storage& open(CheckpointName const&) { std::abort(); }
+    void createCheckpoint(Storage&, CheckpointName const&) {}
+    void deleteCheckpoint(CheckpointName const&) {}
+    std::optional<CheckpointName> latestCheckpointName() const { return std::nullopt; }
+    std::optional<CheckpointName> oldestCheckpointName() const { return std::nullopt; }
+};
+
+using RealGlobalCheckpointBackend = TrivialCheckpointStorage<bcos::executor_v1::StateKey,
+    bcos::executor_v1::StateValue, RealGlobalStateBackendStorage>;
+using RealGlobalStateStorage = bcos::storage2::MultiLayerStorage<RealGlobalStateMutableStorage,
+    void, RealGlobalCheckpointBackend>;
+
+struct StubExecutor
+{
+    template <class Storage>
+    struct ExecuteContext
+    {
+        task::Task<void> prepare() { co_return; }
+        task::Task<void> execute() { co_return; }
+        task::Task<protocol::TransactionReceipt::Ptr> finish() { co_return nullptr; }
+    };
+
+    template <class Storage>
+    task::Task<protocol::TransactionReceipt::Ptr> executeTransaction(Storage&,
+        const protocol::BlockHeader&, const protocol::Transaction&, int,
+        const ledger::LedgerConfig&, bool)
+    {
+        co_return nullptr;
+    }
+
+    template <class Storage>
+    task::Task<ExecuteContext<Storage>> createExecuteContext(Storage&, const protocol::BlockHeader&,
+        const protocol::Transaction&, int, const ledger::LedgerConfig&, bool)
+    {
+        co_return ExecuteContext<Storage>{};
+    }
+};
+
+struct StubScheduler
+{
+    template <class Storage, class Executor>
+    task::Task<std::vector<protocol::TransactionReceipt::Ptr>> executeBlock(Storage&, Executor&,
+        const protocol::BlockHeader&, ::ranges::input_range auto&&, const ledger::LedgerConfig&)
+    {
+        co_return {};
+    }
+};
+
+}  // namespace bcos::engine::eth_test
